CMR14 added
I have added the CMR14, now both mode and MH estimation with prefilter=1 for the CMR14. It should be added back in forecasts, however the posterior mode is still not found.
One gets the regular error message for simple mode compute algorithms, with the full blown MH mode compute, it takes way too long to be able to adress in such a short time.

POSTERIOR KERNEL OPTIMIZATION PROBLEM!
 (minus) the hessian matrix at the "mode" is not positive definite!

*Before going further, I would like to have someone else (with me) to go through the data and how I linked it, as there could be something I did wrong.